+++ owRender/Camera.cpp
@@ -342,7 +342,7 @@ void Camera::CreateRenderable()
 
 
 	// Vertex buffer
-	R_Buffer* __vb = _Render->r.createVertexBuffer(verts.size() * sizeof(vec3), verts.data());
+	R_Buffer* __vb = _Render->r.createVertexBuffer(static_cast<uint32>(verts.size()) * sizeof(vec3), verts.data());
 	__geom = _Render->r.beginCreatingGeometry(_Render->getRenderStorage()->__layout_GxVBF_P);
 	__geom->setGeomVertexParams(__vb, R_DataType::T_FLOAT, 0, 12);
 	__geom->finishCreatingGeometry();
